Check which OS the app is running from
On windows we need to split the files found by "\\",
but not on linux, so I added a conditional to do what it
should based on the os.

This change will also only sort and append the array of md
files if it found any, otherwise it will move on.

[Solves: #7]